Fix SetNumberOfBlades to use size_t instead of non-standard uint
Agent-Logs-Url: https://github.com/kynema/kynema-fmb/sessions/5372f41f-72e4-4942-9053-a0de8e21ba26

+++ src/interfaces/components/controller_builder.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <string>
 #include <string_view>
 
@@ -36,7 +37,7 @@ public:
     }
 
     // Set number of blades
-    ControllerBuilder& SetNumberOfBlades(uint n_blades) {
+    ControllerBuilder& SetNumberOfBlades(size_t n_blades) {
         input.n_blades = n_blades;
         return *this;
     }
